% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stata_available}
\alias{stata_available}
\title{Check Stata Availability}
\usage{
stata_available(error = FALSE)
}
\arguments{
\item{error}{Throw an error if Stata is not found on the system path. Defaults to \code{FALSE}.}
}
\value{
A boolean value, representing whether Stata was found on the system path.
}
\description{
Check Stata Availability
}
\examples{
stata_available()
}
